Update README.md with the current bitcoinkernel branch
I think the documentation for the sub tree is wrong in the readme, I am making a fork of this repo to add some covenants for testing, and the branch i had to update was `KernalApi-48` for `v0.0.22` to get it to compile

`git subtree pull --prefix libbitcoinkernel-sys/bitcoin https://github.com/TheCharlatan/bitcoin kernelApiNode --squash` 

I may have misunderstood something, but i got it to compile now! 

thanks for all the work you have done. i am very excited for this project!